+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.1] - 2025-05-14
+## Fixed
+- Use Hive partition creation time (`createTime`) for scheduling partition cleanup, ensuring accurate expiry timing for both new and existing partitions.
+- Updated workflows to use `Ubuntu 22.04` instead of `Ubuntu 20.04`.
+
 ## [3.6.0] - 2025-03-25
 ## Added
 - Added the `beekeeper.expired.data.table.deletion.enabled` property to separate table deletion into its own configuration item.